Move sudo into its own file
