+++ contracts/marketplace/src/helpers.rs
@@ -1,0 +1,8 @@
+use cosmwasm_std::{Addr, Api, StdResult};
+
+pub fn map_validate(api: &dyn Api, addresses: &[String]) -> StdResult<Vec<Addr>> {
+    addresses
+        .iter()
+        .map(|addr| api.addr_validate(addr))
+        .collect()
+}
