+++ contracts/marketplace/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod error;
 pub mod execute;
+mod helpers;
 pub mod msg;
 mod multitest;
 pub mod query;
 pub mod state;
+pub mod sudo;
 mod unit_tests;
